Remove chart section
